Alex new

+++ src/main/java/de/learny/controller/TestController.java
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.annotation.JsonView;
 
 import de.learny.JsonView.View;
 import de.learny.controller.exception.NotEnoughPermissionsException;
+import de.learny.controller.exception.NotJoinedSubject;
 import de.learny.controller.exception.ResourceNotFoundException;
 import de.learny.dataaccess.TestRepository;
 import de.learny.dataaccess.TestScoreRepository;
@@ -107,18 +108,20 @@ public class TestController {
 
 	@RequestMapping(value = "/{id}/results", method = RequestMethod.POST, consumes = { MediaType.APPLICATION_JSON_VALUE })
 	void turnTest(@PathVariable("id") long id,
-	        @RequestBody Set<Answer> checkedAnswers) {
+	        @RequestBody Set<Question> questions) {
 		Test test = testRepository.findById(id);
 		if (test == null)
 			throw new ResourceNotFoundException("Ein Test mit dieser id existiert nicht");
 		Account loggedInAccount = userToAccountService.getLoggedInAccount();
+		if(!loggedInAccount.getJoinedSubjects().contains(test.getSubject())){
+			throw new NotJoinedSubject("Sie sind nicht in dem Fach angemeldet");
+		}
 		TestScore oldTestScore =  testScoreRepo.findByAccountAndTest(loggedInAccount, test);
 		if(oldTestScore != null){
 			testScoreRepo.delete(oldTestScore);
 		}
-		
-		TestScore testScore = new TestScore(test, loggedInAccount, checkedAnswers);
-		newTestScoreHandler.addNew(testScore);
+		TestScore testScore = new TestScore(test, loggedInAccount, questions);
+		newTestScoreHandler.addNew(testScore, questions);
 	}
 
 	@JsonView(View.Summary.class)
